+++ Sources/Subprocess/AsyncSequence+Additions.swift
@@ -1,0 +1,50 @@
+//
+//  AsyncSequence+Additions.swift
+//  Subprocess
+//
+//  MIT License
+//
+//  Copyright (c) 2023 Jamf Software
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//  SOFTWARE.
+//
+
+import Foundation
+
+extension AsyncSequence {
+    /// Returns a sequence of all the elements.
+    public func sequence() async rethrows -> [Element] {
+        try await reduce(into: [Element]()) { $0.append($1) }
+    }
+}
+
+extension AsyncSequence where Element == UInt8 {
+    /// Returns a `Data` representation.
+    public func data() async rethrows -> Data {
+        Data(try await sequence())
+    }
+    
+    public func string() async rethrows -> String {
+        if #available(macOS 12.0, *) {
+            String(try await characters.sequence())
+        } else {
+            String(decoding: try await data(), as: UTF8.self)
+        }
+    }
+}