+++ Sources/Subprocess/Errors.swift
@@ -28,6 +28,7 @@
 import Foundation
 
 /// Type representing possible errors
+@available(*, deprecated, message: "This type is no longer used with methods supporting Swift Concurrency")
 public enum SubprocessError: Error {
 
     /// The process completed with a non-zero exit code
@@ -47,15 +48,16 @@ public enum SubprocessError: Error {
     case outputStringEncodingError
 }
 
+@available(*, deprecated, message: "This type is no longer used with methods supporting Swift Concurrency")
 extension SubprocessError: LocalizedError {
     public var errorDescription: String? {
         switch self {
         case .exitedWithNonZeroStatus(_, let errorMessage):
             return "\(errorMessage)"
-        case .unexpectedPropertyListObject(_):
+        case .unexpectedPropertyListObject:
             // Ignoring the plist contents parameter as we don't want that in the error message
             return "The property list object could not be cast to expected type"
-        case .unexpectedJSONObject(_):
+        case .unexpectedJSONObject:
             // Ignoring the json contents parameter as we don't want that in the error message
             return "The JSON object could not be cast to expected type"
         case .inputStringEncodingError:
@@ -67,6 +69,7 @@ extension SubprocessError: LocalizedError {
 }
 
 /// Common NSError methods for better interop with Objective-C
+@available(*, deprecated, message: "This type is no longer used with methods supporting Swift Concurrency")
 extension SubprocessError: CustomNSError {
     public var errorCode: Int {
         switch self {